updated spreadsheets and components
I added the x-refs and synonyms (from omiatransbound and omiabreed components) to the master spreadsheets, and I removed the omia components.

This makes things MUCH easier to make changes in the ontology (the information for the same term was scattered across different spreadsheets).
I also removed some columns which we did not need.

FYI: @katiermullen @matentzn